+++ controller/controlplane/controller.go
@@ -25,6 +25,7 @@ import (
 
 	kcfgdataplane "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/dataplane"
 	operatorv1beta1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v1beta1"
+	operatorv2alpha1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v2alpha1"
 	konnectv1alpha2 "github.com/kong/kubernetes-configuration/v2/api/konnect/v1alpha2"
 
 	ctrlconsts "github.com/kong/kong-operator/controller/consts"
@@ -260,6 +261,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 			if err != nil {
 				return ctrl.Result{}, fmt.Errorf("failed to create manager config: %w", err)
 			}
+			mgrCfg.DisableRunningDiagnosticsServer = true
 			if err := r.scheduleInstance(ctx, logger, mgrID, mgrCfg); err != nil {
 				return ctrl.Result{}, err
 			}
@@ -284,6 +286,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 		if err != nil {
 			return ctrl.Result{}, fmt.Errorf("failed to create manager config: %w", err)
 		}
+		mgrCfg.DisableRunningDiagnosticsServer = true
 		hashFromSpec, errSpec := managercfg.Hash(mgrCfg)
 		if errSpec != nil {
 			return ctrl.Result{}, fmt.Errorf("failed to hash ControlPlane config options: %w", errSpec)
@@ -445,6 +448,15 @@ func (r *Reconciler) constructControlPlaneManagerConfigOptions(
 		cfgOpts = append(cfgOpts,
 			WithDataPlaneSyncOptions(*cp.Spec.DataPlaneSync),
 		)
+	}
+
+	if cp.Spec.ConfigDump != nil {
+		if cp.Spec.ConfigDump.State == operatorv2alpha1.ConfigDumpStateEnabled {
+			cfgOpts = append(cfgOpts, WithConfigDumpEnabled(true))
+		}
+		if cp.Spec.ConfigDump.DumpSensitive == operatorv2alpha1.ConfigDumpStateEnabled {
+			cfgOpts = append(cfgOpts, WithSensitiveConfigDumpEnabled(true))
+		}
 	}
 
 	// If the ControlPlane is owned by a Gateway, we set the Gateway to be the only one to reconcile.